Conflicts:
	module/CommonBundle/Component/Lilo/Connection/Http.php
	module/CommonBundle/Component/Lilo/Data/Exception.php
	module/CommonBundle/Controller/Admin/InstallController.php
	module/CommonBundle/Module.php
	module/NewsBundle/Controller/NewsController.php
	module/NewsBundle/Repository/Node/News.php

+++ module/CommonBundle/Controller/AccountController.php
@@ -15,18 +15,25 @@
 namespace CommonBundle\Controller;
 
 use CommonBundle\Component\FlashMessenger\FlashMessage,
+    CommonBundle\Component\PassKit\Pass\Membership,
+    CommonBundle\Component\Util\File\TmpFile,
     CommonBundle\Entity\General\Address,
     CommonBundle\Entity\User\Credential,
+    CommonBundle\Entity\User\Person\Academic,
     CommonBundle\Entity\User\Status\Organization as OrganizationStatus,
     CommonBundle\Entity\User\Status\University as UniversityStatus,
     CommonBundle\Form\Account\Activate as ActivateForm,
+    CommonBundle\Form\Account\Edit as EditForm,
     CudiBundle\Entity\Sale\Booking,
     DateTime,
+    Imagick,
     SecretaryBundle\Entity\Organization\MetaData,
     SecretaryBundle\Entity\Registration,
-    SecretaryBundle\Form\Registration\Edit as EditForm,
     SecretaryBundle\Form\Registration\Subject\Add as SubjectForm,
     Zend\Http\Headers,
+    Zend\File\Transfer\Adapter\Http as FileUpload,
+    Zend\Validator\File\Size as SizeValidator,
+    Zend\Validator\File\IsImage as ImageValidator,
     Zend\View\Model\ViewModel;
 
 /**
@@ -108,6 +115,10 @@ class AccountController extends \SecretaryBundle\Component\Controller\Registrati
             return new ViewModel();
         }
 
+        $registrationEnabled = $this->getEntityManager()
+            ->getRepository('CommonBundle\Entity\General\Config')
+            ->getConfigValue('secretary.registration_enabled') == '1';
+
         $academic = $this->getAuthentication()->getPersonObject();
 
         $studentDomain = $this->getEntityManager()
@@ -129,6 +140,19 @@ class AccountController extends \SecretaryBundle\Component\Controller\Registrati
             $this->getParam('identification')
         );
 
+        $ids = unserialize(
+            $this->getEntityManager()
+                ->getRepository('CommonBundle\Entity\General\Config')
+                ->getConfigValue('secretary.membership_article')
+        );
+
+        $membershipArticles = array();
+        foreach($ids as $organization => $id) {
+            $membershipArticles[$organization] = $this->getEntityManager()
+                ->getRepository('CudiBundle\Entity\Sale\Article')
+                ->findOneById($id);
+        }
+
         if ($this->getRequest()->isPost()) {
             $formData = $this->getRequest()->getPost();
 
@@ -139,7 +163,11 @@ class AccountController extends \SecretaryBundle\Component\Controller\Registrati
                 $formData['become_member'] = isset($formData['become_member']) ? $formData['become_member'] : false;
             $form->setData($formData);
 
-            if ($form->isValid()) {
+            $upload = new FileUpload(array('ignoreNoFile' => true));
+            $upload->addValidator(new SizeValidator(array('max' => '3MB')));
+            $upload->addValidator(new ImageValidator());
+
+            if ($form->isValid() && $upload->isValid()) {
                 $formData = $form->getFormData($formData);
 
                 $universityEmail = $this->_parseUniversityEmail($formData['university_email']);
@@ -197,14 +225,22 @@ class AccountController extends \SecretaryBundle\Component\Controller\Registrati
                     $academic->addUniversityStatus($status);
                 }
 
-                $this->_uploadProfileImage($academic);
+                $this->_uploadProfileImage($upload, $academic);
                 if (isset($formData['organization'])) {
+                    $organization = $this->getEntityManager()
+                        ->getRepository('CommonBundle\Entity\General\Organization')
+                        ->findOneById($formData['organization']);
+
                     $this->_setOrganization(
                         $academic,
                         $this->getCurrentAcademicYear(),
+                        $organization
+                    );
+                } else {
+                    $organization = current(
                         $this->getEntityManager()
                             ->getRepository('CommonBundle\Entity\General\Organization')
-                            ->findOneById($formData['organization'])
+                            ->findAll()
                     );
                 }
 
@@ -215,31 +251,36 @@ class AccountController extends \SecretaryBundle\Component\Controller\Registrati
                 );
 
                 if (null !== $metaData) {
-                    if ($metaData->getTshirtSize() !== null) {
-                        $booking = $this->getEntityManager()
-                            ->getRepository('CudiBundle\Entity\Sale\Booking')
-                            ->findOneAssignedByArticleAndPerson(
-                                $this->getEntityManager()
-                                    ->getRepository('CudiBundle\Entity\Sale\Article')
-                                    ->findOneById($tshirts[$metaData->getTshirtSize()]),
-                                $academic
-                            );
+                    if ($registrationEnabled) {
+                        if (null !== $metaData->getTshirtSize()) {
+                            $booking = $this->getEntityManager()
+                                ->getRepository('CudiBundle\Entity\Sale\Booking')
+                                ->findOneAssignedByArticleAndPerson(
+                                    $this->getEntityManager()
+                                        ->getRepository('CudiBundle\Entity\Sale\Article')
+                                        ->findOneById($tshirts[$metaData->getTshirtSize()]),
+                                    $academic
+                                );
 
-                        if ($booking !== null)
-                            $this->getEntityManager()->remove($booking);
+                            if ($booking !== null)
+                                $this->getEntityManager()->remove($booking);
+                        }
+                        $becomeMember = $metaData->becomeMember() ? true : $formData['become_member'];
+                    } else {
+                        $becomeMember = $metaData->becomeMember();
                     }
-
-                    $becomeMember = $metaData->becomeMember() ? true : $formData['become_member'];
 
                     if ($becomeMember) {
-                        $metaData->setBecomeMember($becomeMember)
-                            ->setReceiveIrReeelAtCudi($formData['irreeel'])
-                            ->setBakskeByMail($formData['bakske'])
-                            ->setTshirtSize($formData['tshirt_size']);
-                    } else {
-                        $metaData->setBakskeByMail($formData['bakske']);
+                        if ($registrationEnabled) {
+                            $metaData->setBecomeMember($becomeMember)
+                                ->setTshirtSize($formData['tshirt_size']);
+                        }
+
+                        $metaData->setReceiveIrReeelAtCudi($formData['irreeel']);
                     }
-                } else {
+
+                    $metaData->setBakskeByMail($formData['bakske']);
+                } elseif ($registrationEnabled) {
                     if ($formData['become_member']) {
                         $metaData = new MetaData(
                             $academic,
@@ -263,25 +304,35 @@ class AccountController extends \SecretaryBundle\Component\Controller\Registrati
                     $this->getEntityManager()->persist($metaData);
                 }
 
-                $membershipArticle = $this->getEntityManager()
-                    ->getRepository('CudiBundle\Entity\Sale\Article')
-                    ->findOneById($this->getEntityManager()
-                        ->getRepository('CommonBundle\Entity\General\Config')
-                        ->getConfigValue('secretary.membership_article')
+                if ($registrationEnabled) {
+                    $membershipArticles = array();
+                    $ids = unserialize(
+                        $this->getEntityManager()
+                            ->getRepository('CommonBundle\Entity\General\Config')
+                            ->getConfigValue('secretary.membership_article')
                     );
 
-                if ($metaData->becomeMember()) {
-                    $this->_bookRegistrationArticles($academic, $formData['tshirt_size']);
-                } else {
-                    $booking = $this->getEntityManager()
-                        ->getRepository('CudiBundle\Entity\Sale\Booking')
-                        ->findOneSoldOrAssignedOrBookedByArticleAndPerson(
-                            $membershipArticle,
-                            $academic
-                        );
+                    foreach($ids as $organizationId => $articleId) {
+                        $membershipArticles[$organizationId] = $this->getEntityManager()
+                            ->getRepository('CudiBundle\Entity\Sale\Article')
+                            ->findOneById($articleId);
+                    }
 
-                    if (null !== $booking)
-                        $this->getEntityManager()->remove($booking);
+                    if ($metaData->becomeMember()) {
+                        $this->_bookRegistrationArticles($academic, $formData['tshirt_size'], $organization, $this->getCurrentAcademicYear());
+                    } else {
+                        foreach($membershipArticles as $membershipArticle) {
+                            $booking = $this->getEntityManager()
+                                ->getRepository('CudiBundle\Entity\Sale\Booking')
+                                ->findOneSoldOrAssignedOrBookedByArticleAndPerson(
+                                    $membershipArticle,
+                                    $academic
+                                );
+
+                            if (null !== $booking)
+                                $this->getEntityManager()->remove($booking);
+                        }
+                    }
                 }
 
                 $academic->activate(
@@ -314,6 +365,8 @@ class AccountController extends \SecretaryBundle\Component\Controller\Registrati
                 $this->_doRedirect();
 
                 return new ViewModel();
+            } else {
+                $form->get('personal')->get('profile')->setMessages($upload->getMessages());
             }
         }
 
@@ -322,6 +375,7 @@ class AccountController extends \SecretaryBundle\Component\Controller\Registrati
                 'form' => $form,
                 'termsAndConditions' => $termsAndConditions,
                 'studentDomain' => $studentDomain,
+                'membershipArticles' => $membershipArticles,
             )
         );
     }
@@ -446,7 +500,6 @@ class AccountController extends \SecretaryBundle\Component\Controller\Registrati
                 $user->setCode(null)
                     ->setCredential(
                         new Credential(
-                            'sha512',
                             $formData['credential']
                         )
                     );
@@ -496,6 +549,33 @@ class AccountController extends \SecretaryBundle\Component\Controller\Registrati
         return new ViewModel(
             array(
                 'data' => $data,
+            )
+        );
+    }
+
+    public function passbookAction()
+    {
+        $pass = new TmpFile();
+        $membership = new Membership(
+            $this->getEntityManager(),
+            $this->getAuthentication()->getPersonObject(),
+            $this->getCurrentAcademicYear(),
+            $pass,
+            'data/images/pass_kit'
+        );
+        $membership->createPass();
+
+        $headers = new Headers();
+        $headers->addHeaders(array(
+            'Content-Disposition' => 'inline; filename="membership.pkpass"',
+            'Content-Type'        => 'application/vnd.apple.pkpass',
+            'Content-Length'      => filesize($pass->getFileName()),
+        ));
+        $this->getResponse()->setHeaders($headers);
+
+        return new ViewModel(
+            array(
+                'data' => $pass->getContent()
             )
         );
     }
@@ -550,6 +630,32 @@ class AccountController extends \SecretaryBundle\Component\Controller\Registrati
             $this->redirect()->toRoute(
                 $this->getParam('return')
             );
+        }
+    }
+
+    private function _uploadProfileImage(FileUpload $upload, Academic $academic)
+    {
+        $filePath = $this->getEntityManager()
+            ->getRepository('CommonBundle\Entity\General\Config')
+            ->getConfigValue('common.profile_path');
+
+        if ($upload->isUploaded()) {
+            $upload->receive();
+
+            $image = new Imagick($upload->getFileName());
+            unlink($upload->getFileName());
+            $image->cropThumbnailImage(320, 240);
+
+            if ($academic->getPhotoPath() != '' || $academic->getPhotoPath() !== null) {
+                $fileName = $academic->getPhotoPath();
+            } else {
+                $fileName = '';
+                do{
+                    $fileName = sha1(uniqid());
+                } while (file_exists($filePath . '/' . $fileName));
+            }
+            $image->writeImage($filePath . '/' . $fileName);
+            $academic->setPhotoPath($fileName);
         }
     }
 }
